+++ rich/box.py
@@ -185,41 +185,41 @@ SQUARE: Box = Box(
 
 MINIMAL: Box = Box(
     """\
-    
+  ╷ 
   │ 
- ─┼ 
+╶─┼╴
   │ 
- ─┼ 
- ─┼ 
+╶─┼╴
+╶─┼╴
   │ 
-    
+  ╵ 
 """
 )
 
 
 MINIMAL_HEAVY_HEAD: Box = Box(
     """\
-    
+  ╷ 
   │ 
- ━┿ 
+╺━┿╸
   │ 
- ─┼ 
- ─┼ 
+╶─┼╴
+╶─┼╴
   │ 
-    
+  ╵ 
 """
 )
 
 MINIMAL_DOUBLE_HEAD: Box = Box(
     """\
-    
+  ╷ 
   │ 
  ═╪ 
   │ 
  ─┼ 
  ─┼ 
   │ 
-    
+  ╵ 
 """
 )
 
@@ -228,10 +228,23 @@ SIMPLE: Box = Box(
     """\
     
     
-────
+╶──╴
     
     
-────
+╶──╴
+    
+    
+"""
+)
+
+SIMPLE_HEAD: Box = Box(
+    """\
+    
+    
+╶──╴
+    
+    
+    
     
     
 """
@@ -254,14 +267,14 @@ SIMPLE_HEAVY: Box = Box(
 
 HORIZONTALS: Box = Box(
     """\
-────
+╶──╴
     
-────
+╶──╴
     
-────
-────
+╶──╴
+╶──╴
     
-────
+╶──╴
 """
 )
 
@@ -399,7 +412,7 @@ if __name__ == "__main__":  # pragma: no cover
         "MINIMAL_HEAVY_HEAD",
         "MINIMAL_DOUBLE_HEAD",
         "SIMPLE",
-        "SIMPLE_HEAVY",
+        "SIMPLE_HEAD" "SIMPLE_HEAVY",
         "HORIZONTALS",
         "ROUNDED",
         "HEAVY",