+++ rich/measure.py
@@ -38,14 +38,15 @@ class Measurement(NamedTuple):
 
     @classmethod
     def get(
-        cls, console: "Console", renderable: "RenderableType", max_width: int
+        cls, console: "Console", renderable: "RenderableType", max_width: int = None
     ) -> "Measurement":
         """Get a measurement for a renderable.
 
         Args:
             console (~rich.console.Console): Console instance.
             renderable (RenderableType): An object that may be rendered with Rich.
-            max_width (int): The maximum width available.
+            max_width (int, optional): The maximum width available, or None to use console.width. 
+                Defaults to None.
 
         Raises:
             errors.NotRenderableError: If the object is not renderable.
@@ -54,6 +55,7 @@ class Measurement(NamedTuple):
             Measurement: Measurement object containing range of character widths required to render the object.
         """
 
+        _max_width = console.width if max_width is None else max_width
         if isinstance(renderable, str):
             renderable = console.render_str(renderable)
 
@@ -61,12 +63,12 @@ class Measurement(NamedTuple):
         if is_renderable(renderable):
             get_console_width = getattr(renderable, "__rich_measure__", None)
             if get_console_width is not None:
-                render_width = get_console_width(console, max_width).with_maximum(
-                    max_width
+                render_width = get_console_width(console, _max_width).with_maximum(
+                    _max_width
                 )
                 return render_width.normalize()
             else:
-                return Measurement(1, max_width)
+                return Measurement(1, _max_width)
         else:
             raise errors.NotRenderableError(
                 f"Unable to get render width for {renderable!r}; "