+++ rich/segment.py
@@ -70,7 +70,7 @@ class Segment(NamedTuple):
 
     @classmethod
     def filter_control(
-        self, segments: Iterable["Segment"], is_control=False
+        cls, segments: Iterable["Segment"], is_control=False
     ) -> Iterable["Segment"]:
         """Filter segments by ``is_control`` attribute.